fix(web): use vite routing conventions and revert lockfile hoist (#2)
- Renamed Next.js style `app/agents/[id]` to Vite/React Router style
  `app/agents/_id`.
- Updated router imports in `main.tsx`.
- Reverted unintentional modifications to root package.json and
  package-lock.json that hoisted web dependencies to the workspace root.
