+++ web/src/main.tsx
@@ -14,8 +14,8 @@ import { ErrorBoundary } from '@/components/ErrorBoundary';
 import DashboardPage from '@/app/dashboard/page';
 import ChatPage from '@/app/chat/page';
 import AgentsPage from '@/app/agents/page';
-import AgentDetailPage from '@/app/agents/[id]/page';
-import AgentEditPage from '@/app/agents/[id]/edit/page';
+import AgentDetailPage from '@/app/agents/_id/page';
+import AgentEditPage from '@/app/agents/_id/edit/page';
 import AgentNewPage from '@/app/agents/new/page';
 import CirclesPage from '@/pages/CirclesPage';
 import CircleDetailPage from '@/pages/CircleDetailPage';
